# Lua Scripting Allocation, Performance, and Correctness Improvements